feat: Clerk Authentication
