+++ app/(routes)/admin/_actions.ts
@@ -1,0 +1,23 @@
+"use server";
+
+import { checkRole } from "@/utils/roles";
+import { clerkClient } from "@clerk/nextjs/server";
+
+export async function setRole(formData: FormData) {
+  // Check that the user trying to set the role is an admin
+  if (!checkRole("admin")) {
+    return { message: "Not Authorized" };
+  }
+
+  try {
+    const res = await clerkClient.users.updateUser(
+      formData.get("id") as string,
+      {
+        publicMetadata: { role: formData.get("role") },
+      }
+    );
+    return { message: res.publicMetadata };
+  } catch (err) {
+    return { message: err };
+  }
+}
