+++ types/globals.d.ts
@@ -1,0 +1,12 @@
+export {};
+
+// Create a type for the roles
+export type Roles = "admin" | "kitchen" | "user";
+
+declare global {
+  interface CustomJwtSessionClaims {
+    metadata: {
+      role?: Roles;
+    };
+  }
+}
