+++ utils/roles.ts
@@ -1,0 +1,8 @@
+import { Roles } from "@/types/globals";
+import { auth } from "@clerk/nextjs/server";
+
+export const checkRole = (role: Roles) => {
+  const { sessionClaims } = auth();
+
+  return sessionClaims?.metadata.role === role;
+};
